Bump version -> v0.6.1
The last publish failed because pyproject.toml was out-of-sync before
the last bump. Now that everything is in sync, publish should work.

As a side note, we can't just re-tag because we've already published a
Zenodo release. We should decouple the Zenodo release from the package
publish. I.e. package publish is triggered by tag push, then Zenodo
is triggered by GH Release on the same tag, post- successful publish.