Split the README into a GitHub README and doc page
Previously, the README file was symlinked into our docs. This led to a
problem where we could only write links that worked within one of the
two systems. Docs links need to be relative to the symlink to render
correctly, and GitHub README links need to be relative to the original
file to work correctly.