Allow pip install to run when conda env is active
If a user has either the `PIP_REQUIRE_VENV` or `PIP_REQUIRE_VIRTUALENV`
environment variable set to a "truthy" value, and no virtual environment
is active, `pip install` will fail with a message indicating as such.
However, when a conda environment is active, we want to allow
`pip install` to run without an active virtual environment because it
will install packages within the active conda environment.